refactor : 수익 결정하는 책임 BetAmount에서 Incomes로 전달
